code cleanup

+++ src/main/java/gremlin/patches/eventpatches/CursedTomePatch.java
@@ -17,7 +17,7 @@ public class CursedTomePatch {
     {
         public static void Postfix(CursedTome __instance) {
             if (AbstractDungeon.player instanceof GremlinCharacter) {
-                int dmg = (int) ReflectionHacks.getPrivate(__instance, CursedTome.class, "finalDmg");
+                int dmg = ReflectionHacks.getPrivate(__instance, CursedTome.class, "finalDmg");
                 dmg = (dmg + 4) / 5; // Divided by 5 rounded up
                 ReflectionHacks.setPrivate(__instance, CursedTome.class, "finalDmg", dmg);
             }
@@ -32,7 +32,7 @@ public class CursedTomePatch {
         )
         public static void Insert(CursedTome __instance, int buttonPressed){
             if (AbstractDungeon.player instanceof GremlinCharacter) {
-                int dmg = (int) ReflectionHacks.getPrivate(__instance, CursedTome.class, "finalDmg");
+                int dmg = ReflectionHacks.getPrivate(__instance, CursedTome.class, "finalDmg");
                 __instance.imageEventText.updateDialogOption(0, strings.OPTIONS[0] + dmg + CursedTome.OPTIONS[6]);
             }
         }
@@ -46,7 +46,7 @@ public class CursedTomePatch {
         )
         public static void Insert(CursedTome __instance, int buttonPressed){
             if (AbstractDungeon.player instanceof GremlinCharacter) {
-                int dmg = (int) ReflectionHacks.getPrivate(__instance, CursedTome.class, "finalDmg");
+                int dmg = ReflectionHacks.getPrivate(__instance, CursedTome.class, "finalDmg");
                 ((GremlinCharacter) AbstractDungeon.player).damageGremlins(dmg);
             }
         }
